Migrate to Minimal APIs
- https://docs.microsoft.com/aspnet/core/fundamentals/minimal-apis
