+++ samples/SampleWebApp/appsettings.json
@@ -3,11 +3,10 @@
     "System.Data.SQLite": "Data Source=|DataDirectory|/database.sqlite;"
   },
   "Logging": {
-    "IncludeScopes": "false",
     "LogLevel": {
-      "Default": "Debug",
-      "System": "Information",
-      "Microsoft": "Information"
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
+      "NHibernate": "Debug"
     }
   }
 }
